Métodos da classe Grafos implementados

+++ Aresta.py
@@ -2,7 +2,7 @@ from Vertice import Vertice
 
 
 class Aresta:
-    def __init__(self, u: Vertice, v: Vertice, peso: int):
+    def __init__(self, u: Vertice, v: Vertice, peso: float):
         self.u = u
         self.v = v
         self.peso = peso
